split errors proc dir
