+++ src/errors/mod.rs
@@ -1,0 +1,2 @@
+// errors パース時に発生したエラー処理
+pub mod parser_errors;
