+++ src/parser/mod.rs
@@ -5,6 +5,3 @@ pub mod core_parser;
 // core_parser.parser for *
 pub mod expr_parser;
 pub mod state_parser;
-
-// errors パース時に発生したエラー処理
-pub mod parser_errors;
