+++ src/token/func.rs
@@ -1,5 +1,5 @@
 use crate::abs::ast::*;
-use crate::parser::parser_errors::ParserError;
+use crate::errors::parser_errors::ParserError;
 use crate::token::paren_block::ParenBlockBranch;
 
 /// # FuncBranch
