+++ src/token/paren_block.rs
@@ -1,6 +1,6 @@
 use crate::abs::ast::*;
+use crate::errors::parser_errors::ParserError;
 use crate::parser::expr_parser::ExprParser;
-use crate::parser::parser_errors::ParserError;
 
 /// #ParenBlockBranch
 /// `()`を使用したプログラムにおけるデータを格納するstruct
